Функции random перенесены в файл utils.js
